engine/requirements.txt
strawberry-graphql
uvicorn[standard]
pydantic
python-dateutil